Removing FLTK usage from libstage
I've separated gui-specific code to libstage-gui library, so base simulation does not need fltk anymore.

Now stage compiles two executables: stage and stage-headless. Headless variant does not needs FLTK

Fixed most of compiler errors. Linker to be fixed
