+++ libstage/option.cc
@@ -8,7 +8,7 @@ using namespace Stg;
 Option::Option(const std::string &n, const std::string &tok, const std::string &key, bool v,
                World *world)
     : optName(n), value(v), wf_token(tok), shortcut(key), menu(NULL), menuIndex(0), menuCb(NULL),
-      menuCbWidget(NULL), _world(world), htname(n)
+			callbackData(NULL), _world(world), htname(n)
 {
   /* do nothing */
 }
@@ -30,28 +30,10 @@ void Option::Save(Worldfile *wf, int section)
   wf->WriteInt(section, wf_token.c_str(), value);
 }
 
-void Option::toggleCb(Fl_Widget *, void *p)
-{
-  // Fl_Menu_* menu = static_cast<Fl_Menu_*>( w );
-  Option *opt = static_cast<Option *>(p);
-  opt->invert();
-  if (opt->menuCb)
-    opt->menuCb(opt->menuCbWidget, opt);
-}
-
-void Option::menuCallback(Fl_Callback *cb, Fl_Widget *w)
+void Option::setOptionCallback(Option::Callback cb, void *data)
 {
   menuCb = cb;
-  menuCbWidget = w;
-}
-
-void Option::createMenuItem(Fl_Menu_Bar *m, std::string path)
-{
-  menu = m;
-  path = path + "/" + optName;
-  // create a menu item and save its index
-  menuIndex = menu->add(path.c_str(), shortcut.c_str(), toggleCb, this,
-                        FL_MENU_TOGGLE | (value ? FL_MENU_VALUE : 0));
+  this->callbackData = data;
 }
 
 void Option::set(bool val)
@@ -64,11 +46,13 @@ void Option::set(bool val)
   }
 
   if (_world) {
+#ifdef TEST_LATER
     WorldGui *wg = dynamic_cast<WorldGui *>(_world);
     if (wg == NULL)
       return;
     Canvas *canvas = wg->GetCanvas();
-    canvas->invalidate();
+    canvas->setInvalidate();
     canvas->redraw();
+#endif
   }
 }
