+++ libstage/worldgui.cc
@@ -158,6 +158,7 @@ then select the option from the menu again to stop.
 #include "options_dlg.hh"
 #include "region.hh"
 #include "worldfile.hh"
+#include "world_gui.hh"
 
 using namespace Stg;
 
@@ -189,7 +190,7 @@ static const char *MoreHelpText =
     "(requires Doxygen and supporting programs to be installed first).\n";
 
 WorldGui::WorldGui(int width, int height, const char *caption)
-    : Fl_Window(width, height, NULL), canvas(new Canvas(this, 0, 30, width, height - 30)),
+    : Fl_Window(width, height, NULL), canvas(new CanvasFLTK(this, 0, 30, width, height - 30)),
       drawOptions(), fileMan(new FileManager()), interval_log(), speedup(1.0), // real time
       confirm_on_quit(true), mbar(new Fl_Menu_Bar(0, 0, width, 30)), oDlg(NULL), pause_time(false),
       real_time_interval(sim_interval), real_time_now(RealTimeNow()),
@@ -453,36 +454,6 @@ std::string WorldGui::EnergyString() const
            (PowerPack::global_dissipated / (sim_time / 1e6)) / 1e3);
   
   return std::string(str);
-}
-
-void WorldGui::DrawOccupancy() const
-{
-  // 	int count=0;
-  //   FOR_EACH( it, superregions )
-  // 		printf( "sr %d [%d,%d]  %p\n", count++, it->first.x,
-  // it->first.y,
-  // it->second );
-  // 	printf( "done\n" );
-
-  //  unsigned int layer( updates % 2 );
-
-  FOR_EACH (it, superregions)
-    it->second->DrawOccupancy();
-
-  // 	 {
-
-  // it->second->DrawOccupancy(0);
-  //    it->second->DrawOccupancy(1);
-
-  //	 }
-}
-
-void WorldGui::DrawVoxels() const
-{
-  unsigned int layer(updates % 2);
-
-  FOR_EACH (it, superregions)
-    it->second->DrawVoxels(layer);
 }
 
 void WorldGui::windowCb(Fl_Widget *, WorldGui *wg)
@@ -854,12 +825,6 @@ bool WorldGui::closeWindowQuery()
     // nothing is loaded or confirmation not wanted, just quit
     return true;
   }
-}
-
-void WorldGui::DrawBoundingBoxTree()
-{
-  FOR_EACH (it, World::children)
-    (*it)->DrawBoundingBoxTree();
 }
 
 void WorldGui::PushColor(Color col)
